Added posts, fixed bugs
Finished Volume form posted
Revamped the Hodge Star post
Fixed the issue with the pdf on mobile devices

Next steps are that I need to finish the Hodge star post and to add a new one on solutions to schrodinger equations and other such things. Maybe add a post on SU(2) representations on SO(3) making it a double cover and how to generalize to form the spin groups from clifford algebras.